Improve performance of API data transformation
Related to #1010
