+++ .changeset/unlucky-kiwis-brush.md
@@ -1,0 +1,5 @@
+---
+"alephium-desktop-wallet": patch
+---
+
+Improve performance
