+++ apps/desktop-wallet/src/api/apiDataHooks/utils/useFetchWalletBalancesAlph.ts
@@ -1,0 +1,50 @@
+/*
+Copyright 2018 - 2024 The Alephium Authors
+This file is part of the alephium project.
+
+The library is free software: you can redistribute it and/or modify
+it under the terms of the GNU Lesser General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+The library is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+GNU Lesser General Public License for more details.
+
+You should have received a copy of the GNU Lesser General Public License
+along with the library. If not, see <http://www.gnu.org/licenses/>.
+*/
+
+import { useQueries, UseQueryResult } from '@tanstack/react-query'
+
+import { addressAlphBalancesQuery, AddressAlphBalancesQueryFnData } from '@/api/queries/addressQueries'
+import { useAppSelector } from '@/hooks/redux'
+import { useUnsortedAddressesHashes } from '@/hooks/useAddresses'
+import { selectCurrentlyOnlineNetworkId } from '@/storage/network/networkSelectors'
+
+const useFetchWalletBalancesAlph = <T>(
+  combine: (results: UseQueryResult<AddressAlphBalancesQueryFnData>[]) => {
+    data: T
+    isLoading: boolean
+    isFetching?: boolean
+    error?: boolean
+  }
+) => {
+  const networkId = useAppSelector(selectCurrentlyOnlineNetworkId)
+  const allAddressHashes = useUnsortedAddressesHashes()
+
+  const { data, isLoading, isFetching, error } = useQueries({
+    queries: allAddressHashes.map((addressHash) => addressAlphBalancesQuery({ addressHash, networkId })),
+    combine
+  })
+
+  return {
+    data,
+    isLoading,
+    isFetching,
+    error
+  }
+}
+
+export default useFetchWalletBalancesAlph
