+++ apps/desktop-wallet/src/api/apiDataHooks/wallet/useFetchWalletWorthAlph.ts
@@ -21,7 +21,7 @@ import { ALPH } from '@alephium/token-list'
 import { useMemo } from 'react'
 
 import { useFetchTokenPrice } from '@/api/apiDataHooks/market/useFetchTokenPrices'
-import { useFetchWalletBalancesAlphArray } from '@/api/apiDataHooks/wallet/useFetchWalletBalancesAlph'
+import useFetchWalletBalancesAlphArray from '@/api/apiDataHooks/wallet/useFetchWalletBalancesAlphArray'
 
 const useFetchWalletWorthAlph = () => {
   const { data: alphBalances, isLoading: isLoadingAlphBalances } = useFetchWalletBalancesAlphArray()
