+++ apps/desktop-wallet/src/index.tsx
@@ -28,6 +28,7 @@ import { Provider } from 'react-redux'
 import { HashRouter as Router } from 'react-router-dom'
 import { StyleSheetManager } from 'styled-components'
 
+import ApiContextProvider from '@/api/apiContext'
 import { PersistQueryClientContextProvider } from '@/api/persistQueryClientContext'
 import App from '@/App'
 import Tooltips from '@/components/Tooltips'
@@ -48,7 +49,9 @@ ReactDOM.render(
           <Suspense fallback="loading">
             <StyleSheetManager shouldForwardProp={shouldForwardProp}>
               <PersistQueryClientContextProvider>
-                <App />
+                <ApiContextProvider>
+                  <App />
+                </ApiContextProvider>
                 <ReactQueryDevtools initialIsOpen={false} />
               </PersistQueryClientContextProvider>
               <Tooltips />
